+++ imod/mf6/mf6_wel_adapter.py
@@ -15,8 +15,11 @@ consisting of layer, row, and column, closely resembling input for Modflow6.
 from typing import Any, Dict, Optional
 
 import numpy as np
+import xarray as xr
+from numpy.typing import NDArray
 
 from imod.mf6.boundary_condition import BoundaryCondition
+from imod.mf6.interfaces.ipackage import IPackage
 from imod.schemata import DTypeSchema
 
 # FUTURE: There was an idea to autogenerate modflow 6 adapters.
@@ -24,7 +27,95 @@ from imod.schemata import DTypeSchema
 # https://github.com/Deltares/xugrid/blob/main/xugrid/core/wrap.py#L90
 
 
-class Mf6Wel(BoundaryCondition):
+def cellid_from_arrays__structured(
+    layer: NDArray[np.int_], row: NDArray[np.int_], column: NDArray[np.int_]
+):
+    """
+    Create DataArray of cellids for structured indices.
+
+    Parameters
+    ----------
+    layer: numpy.array
+        1D array of integers for layer
+    row: numpy.array
+        1D array of integers for row
+    column: numpy.array
+        1D array of integers for column
+
+    Returns
+    -------
+    cellid: xr.DataArray
+        2D DataArray with ``ncellid`` rows and 3 columns.
+    """
+    return _cellid_from_kwargs(layer=layer, row=row, column=column)
+
+
+def cellid_from_arrays__unstructured(layer: NDArray[np.int_], cell2d: NDArray[np.int_]):
+    """
+    Create DataArray of cellids for unstructured indices.
+
+    Parameters
+    ----------
+    layer: numpy.array
+        1D array of integers for layer
+    cell2d: numpy.array
+        1D array of integers for cell2d indices
+
+    Returns
+    -------
+    cellid: xr.DataArray
+        2D DataArray with ``ncellid`` rows and 2 columns.
+    """
+    return _cellid_from_kwargs(layer=layer, cell2d=cell2d)
+
+
+def _cellid_from_kwargs(**kwargs):
+    dim_cellid_coord = list(kwargs.keys())
+    arr_ls = list(kwargs.values())
+    index = np.arange(len(arr_ls[0]))
+    indices_ls = [
+        xr.DataArray(ind_arr, coords={"index": index}, dims=("index",))
+        for ind_arr in arr_ls
+    ]
+    return concat_indices_to_cellid(indices_ls, dim_cellid_coord)
+
+
+def concat_indices_to_cellid(
+    indices_ls: list[xr.DataArray], dim_cellid_coord: list[str]
+) -> xr.DataArray:
+    """
+    Create DataArray of cellids from list of one-dimensional DataArrays with
+    indices.
+
+    Parameters
+    ----------
+    indices_ls: list of xr.DataArrays
+        List of one-dimensional DataArrays with indices. For structured grids,
+        these are the layer, rows, columns. For unstructured grids, these are
+        layer, cell2d.
+    dim_cellid_coord: list of strings
+        List of coordinate names, which are assigned to the ``"dim_cellid"``
+        dimension.
+
+    Returns
+    -------
+    cellid: xr.DataArray
+        2D DataArray with ``ncellid`` rows and 3 to 2 columns, depending
+        on whether on a structured or unstructured grid.
+    """
+    cellid = xr.concat(indices_ls, dim="dim_cellid")
+    # Rename generic dimension name "index" to ncellid.
+    cellid = cellid.rename(index="ncellid")
+    # Put dimensions in right order after concatenation.
+    cellid = cellid.transpose("ncellid", "dim_cellid")
+    # Assign extra coordinate names.
+    coords = {
+        "dim_cellid": dim_cellid_coord,
+    }
+    return cellid.assign_coords(coords=coords)
+
+
+class Mf6Wel(BoundaryCondition, IPackage):
     """
     Package resembling input for Modflow 6 List Input. This class has
     methods for the modflow 6 wel packages with time component.
@@ -86,7 +177,7 @@ class Mf6Wel(BoundaryCondition):
             dsvar[var] = ds[var]
         arrdict["var_values"] = dsvar
 
-        arrdict["cellid_names"] = ds.coords["nmax_cellid"].values
+        arrdict["cellid_names"] = ds.coords["dim_cellid"].values
         arrdict["nrow"] = ds.coords["ncellid"].size
         arrdict["cellid"] = ds["cellid"]
 
@@ -100,7 +191,7 @@ class Mf6Wel(BoundaryCondition):
         # Initialize the structured array
         recarr = np.empty(arrdict["nrow"], dtype=sparse_dtype)
         for cellid_name in arrdict["cellid_names"]:
-            recarr[cellid_name] = arrdict["cellid"].sel(nmax_cellid=cellid_name).values
+            recarr[cellid_name] = arrdict["cellid"].sel(dim_cellid=cellid_name).values
 
         for var in arrdict["data_vars"]:
             recarr[var] = arrdict["var_values"][var]